# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"@project.name@","description":"@project.description@","contact":{"name":"Bruno Emanuel Toledo","email":"bruno@gmail.com"},"version":"@project.version@"},"servers":[{"url":"http://localhost:8080","description":"@project.description@"}],"paths":{"/players":{"post":{"tags":["player-controller"],"summary":"crear un player","description":"retorna el player creado con su id.Si el player existe con el username o email, retorna un error 404, adicionalmente, el mail debe ser valido y el password debe ser mayor a 8 caracteres y contener un numero,tambien tendra que tener un caracter en mayuscula, y un caracter especial","operationId":"savePlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Player"}}},"required":true},"responses":{"500":{"description":"error en el servidor","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"Se creo el player","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"404":{"description":"el nombre o el email ya existen","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/players/{id}":{"get":{"tags":["player-controller"],"summary":"trae un player por id","description":"retorna el jugador si exite el id. Sino existe retorna un erro 404","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"no encontrado","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"error en el servidor","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"operacion exitosa","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/players/allPlayers":{"get":{"tags":["player-controller"],"operationId":"getAllPlayers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Player"}}}}}}}},"/pico":{"get":{"tags":["picoy-pala-controller"],"summary":"chequea el estado de salud de nuestra app","description":"si la app responde pala esta sana","operationId":"pala","responses":{"200":{"description":"Operacion exitosa","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Error interno en el servidor","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"404":{"description":"No encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}}},"components":{"schemas":{"Player":{"required":["email","password","userName"],"type":"object","properties":{"id":{"title":"player Id","type":"integer","description":"el id del player","format":"int64","example":1},"userName":{"title":"nombre del player","type":"string","description":"el nombre del player","example":"Bruno"},"password":{"title":"password del player","type":"string","description":"el password del player","example":"Password3#"},"email":{"title":"player email","type":"string","description":"el mail del player","example":"email@email.com"},"avatar":{"title":"player avatar","type":"string","description":"avatar player","nullable":true,"example":"avatar player"},"lastLoginDate":{"title":"player last login","type":"string","description":"the player last login","format":"date-time","nullable":true}}},"ErrorApi":{"type":"object","properties":{"timeStamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}